bugs/provision sealed sector file (#74)
* fix(sealed-file): ensure that sealed sector-file exists before pre-commit

* fix(api): ensure seed and ticket are exposed through API

* fix(precommit): expose pre-commit output as return value

* fix(tests): don't run transition test twice

* fix(tests): don't use FFISealedStatusMetadata

* feat(lock): update lockfile

* fix(pre-commit): pre-commit response should include commitments

* temp(test): minimal test for leak debug

* feat(toolchain): upgrade toolchain to what rust-fil-proofs will use

* temp(test): minimal test for leak debug

* fix(clippy): updates for clippy after bumping toolchain version

* fix(bindgen): update to newest version of bindgen + cbindgen

* feat(workers): configurable worker thread count

* fix(upstream): roll back to pre-GPU version of rust-fil-proofs which doesn't panic

* fix(tests): drop debug-only test

* fix(toolchain): downgrade to bls-signatures compat version

+++ sector-builder/src/worker.rs
@@ -15,7 +15,7 @@ const FATAL_NOLOCK: &str = "error acquiring task lock";
 const FATAL_RCVTSK: &str = "error receiving task";
 
 pub struct Worker {
-    pub id: usize,
+    pub id: u8,
     pub thread: Option<thread::JoinHandle<()>>,
 }
 
@@ -111,7 +111,7 @@ pub enum WorkerTask {
 
 impl Worker {
     pub fn start(
-        id: usize,
+        id: u8,
         seal_task_rx: Arc<Mutex<mpsc::Receiver<WorkerTask>>>,
         prover_id: [u8; 32],
     ) -> Worker {
